+++ src/components/FeedView/Comments/CommentContents/CommentContents.jsx
@@ -4,6 +4,7 @@ import { toast } from 'react-toastify';
 import { user } from 'api/queries/user';
 import { comments } from 'api/queries/comment';
 import { EditDeleteButton, Icon, ProfileImage } from 'components/Common';
+import { EditForm } from '../EditForm/EditForm';
 import { ReactComponent as Comment } from 'assets/icons/comment.svg';
 import { dateFormat } from 'utils/dateConverter';
 import * as S from './styles';
@@ -20,31 +21,8 @@ export const CommentContents = ({
   const { author, content, created_at, comment_id } = contents;
 
   const [isEditing, setIsEditing] = useState(false);
-  const [editValue, setEditValue] = useState(content);
 
   const queryClient = useQueryClient();
-
-  // 댓글 수정
-  const { mutate: editCommentMutate } = useMutation(comments.edit, {
-    onSuccess: () => {
-      toast.success('댓글이 수정되었습니다!');
-      queryClient.invalidateQueries('comments');
-    },
-    onError: error => {
-      console.log(error.message);
-      toast.error('댓글 수정에 실패했습니다. 다시 시도해주세요.');
-    }
-  });
-
-  const handleChange = e => {
-    setEditValue(e.target.value);
-  };
-
-  const handleEditComment = e => {
-    e.preventDefault();
-    editCommentMutate([{ commentId: comment_id }, { content: editValue }]);
-    setIsEditing(prev => !prev);
-  };
 
   // 댓글 삭제
   const { mutate: deleteCommentMutate } = useMutation(comments.delete, {
@@ -88,10 +66,11 @@ export const CommentContents = ({
 
       <S.CommentText>
         {isEditing ? (
-          <S.EditingForm onSubmit={handleEditComment}>
-            <S.EditInput value={editValue} onChange={handleChange} />
-            <button onClick={handleEditComment}>완료</button>
-          </S.EditingForm>
+          <EditForm
+            content={content}
+            commentId={comment_id}
+            setIsEditing={setIsEditing}
+          />
         ) : (
           <S.TextBox>{content}</S.TextBox>
         )}
