+++ src/components/FeedView/Comments/CommentContents/styles.js
@@ -39,25 +39,6 @@ export const TextBox = styled.p`
   word-break: break-all;
 `;
 
-export const EditingForm = styled.form`
-  display: flex;
-  button {
-    padding: 4px 6px;
-    margin-left: 4px;
-    font-size: 13px;
-    line-height: 19.5px;
-    border: 1px solid #eaeaea;
-    border-radius: 4px;
-  }
-`;
-
-export const EditInput = styled.input`
-  flex: 1;
-  padding: 4px 6px;
-  border: 1px solid #eaeaea;
-  border-radius: 4px;
-`;
-
 export const CommentFooter = styled.div`
   display: flex;
   justify-content: space-between;
