+++ src/components/FeedView/Comments/EditForm/styles.js
@@ -1,0 +1,20 @@
+import styled from 'styled-components';
+
+export const EditingForm = styled.form`
+  display: flex;
+  button {
+    padding: 4px 6px;
+    margin-left: 4px;
+    font-size: 13px;
+    line-height: 19.5px;
+    border: 1px solid #eaeaea;
+    border-radius: 4px;
+  }
+`;
+
+export const EditInput = styled.input`
+  flex: 1;
+  padding: 4px 6px;
+  border: 1px solid #eaeaea;
+  border-radius: 4px;
+`;
