+++ src/components/FeedView/index.js
@@ -1,4 +1,5 @@
 export { ImageSlider } from './ImageSlider/ImageSlider';
 export { CommentBody } from './Comments/CommentBody/CommentBody';
 export { CommentForm } from './Comments/CommentForm/CommentForm';
+export { EditForm } from './Comments/EditForm/EditForm';
 export { ReplyForm } from './Comments/ReplyForm/ReplyForm';
